+++ src/dactyl_keyboard/cots.clj
@@ -82,13 +82,15 @@
 
 (def switch-facts
   "Form factors of switches for the purpose of cutting holes into key mounting
-  plates. The dmote-keycap libraries models switches as such in more detail
+  plates. The dmote-keycap library models switches as such in more detail
   for the contrasting purpose of modeling caps."
   {:alps {:hole {:x 15.5,  :y 12.6}
           :foot {:x 17.25, :y 14.25}
+          :height {:above-plate 7.6, :into-plate 4.5}
           :description "ALPS-style, including Matias"}
    :mx   {:hole {:x 14,    :y 14}
           :foot {:x 15.5,  :y 15.5}
+          :height {:above-plate 5, :into-plate 5}
           :description "Cherry MX style"}})
 
 (defn support-list